상품 주문시 남은 재고 표시

+++ src/main/java/com/spharosacademy/project/SSGBack/product/option/entity/OptionList.java
@@ -19,7 +19,7 @@ public class OptionList {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private int stock;
+    private float stock;
 
     @ManyToOne
     @JsonBackReference
